Finished, Made back Office Controller and Service with it's Dependency Injection For Changing LoanStatus.

+++ LoanProject.Services/Implementations/LoanService.cs
@@ -1,31 +1,20 @@
 ﻿using AutoMapper;
-using Azure.Core;
 using LoanProject.Data.Models;
 using LoanProject.Data.Models.Enums;
 using LoanProject.Repository.Abstractions;
 using LoanProject.Services.Abstractions;
-using LoanProject.Services.Models;
-using LoanProject.Services.Models.Enum;
 using LoanProject.Services.Models.Loan.LoanServiceRequest;
 using LoanProject.Services.Models.Loan.LoanServiceResponses;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
+using Microsoft.AspNetCore.Http;
 
 namespace LoanProject.Services.Implementations
 {
     public class LoanService : ILoanService
     {
-        private readonly IPasswordHasher _hasher;
-        private readonly IBaseRepository<Loan> _repo;
         private readonly ILoanRepository _loanRepo;
         private readonly IMapper _mapper;
-        public LoanService(IPasswordHasher hasher, IJwtService jwtService, IMapper mapper, IBaseRepository<Loan> repo, ILoanRepository loanRepo)
+        public LoanService(IMapper mapper, ILoanRepository loanRepo)
         {
-            _hasher = hasher;
-            _repo = repo;
             _mapper = mapper;
             _loanRepo = loanRepo;
         }
@@ -34,51 +23,82 @@ namespace LoanProject.Services.Implementations
         {
             var insertModel = _mapper.Map<Loan>(request);
             insertModel.UserId = userId;
-            await _repo.AddAsync(insertModel);
+            await _loanRepo.AddAsync(insertModel);
 
             return new TakeLoanResponse
             {
                 Message = "Loan request is sent",
-                Status = LoanStatus.Pending,
-                StatusCode = StatusCodes.Success
+                Status = LoanStatus.Forwarded.ToString(),
+                StatusCode = StatusCodes.Status200OK
             };
         }
 
         public async Task<IEnumerable<Loan>> GetLoans()
         {
-            return await _repo.GetAllAsync();
+            return await _loanRepo.GetAllAsync();
         }
 
         public async Task<IEnumerable<Loan>> GetLoanByUserIdAsync(int userId)
         {
             return await _loanRepo.GetLoansByUserIdAsync(x => x.UserId == userId);
         }
+
         public async Task<EditLoanResponse> EditLoanInfoAsync(UpdateLoanRequest request, int loanId)
         {
             var loan = await _loanRepo.GetByIdAsync(loanId);
-            loan.loanType = request.loanType;
-            loan.Amount = request.Amount;
-            loan.TermInMonths = request.TermInMonths;
 
-            _loanRepo.Update(loan);
+            if (loan != null)
+            {
+                if (loan.LoanStatus == LoanStatus.Pending)
+                {
+                    loan.loanType = request.loanType;
+                    loan.Amount = request.Amount;
+                    loan.TermInMonths = request.TermInMonths;
+
+                    _loanRepo.Update(loan);
+                    return new EditLoanResponse
+                    {
+                        Message = "Loan is Updated !",
+                        StatusCode = StatusCodes.Status200OK,
+                        UserId = loan.UserId
+                    };
+                }
+                return new EditLoanResponse
+                {
+                    Message = "Your loan has already been assigned a status, so it cannot be edited",
+                    StatusCode = StatusCodes.Status200OK,//.
+                    UserId = loan.UserId
+                };
+            }
+
             return new EditLoanResponse
             {
-                Message = "Loan is Updated !",
-                StatusCode = StatusCodes.Success,
-                UserId = loan.UserId
+                Message = "Loan have not found with this LoanId ! ",
+                StatusCode = StatusCodes.Status400BadRequest,
+                UserId = 0
             };
-
         }
 
-        public async Task<CancelLoanResponse> CancelLoan(int LoanId)
+        public async Task<CancelLoanResponse> CancelLoan(int loanId)
         {
-            var loan = await _loanRepo.GetByIdAsync(LoanId);
-            _loanRepo.Remove(loan);
+            var loan = await _loanRepo.GetByIdAsync(loanId);
+
+            if (loan != null)
+            {
+                _loanRepo.Remove(loan);
+                return new CancelLoanResponse
+                {
+                    LoanId = loanId,
+                    Message = "Loan request is cancelled",
+                    StatusCode = StatusCodes.Status200OK
+                };
+            }
+
             return new CancelLoanResponse
             {
-                LoanId = LoanId,
-                Message = "Loan request is cancelled",
-                StatusCode = StatusCodes.Success
+                LoanId = loanId,
+                Message = "Loan could not be found with this LoanId",
+                StatusCode = StatusCodes.Status400BadRequest
             };
         }
     }
